gst/va: move AV1 encoder test classes to lib
